requirements.txt
django
mysqlclient
django_rest_framework
pillow
drf-extra-fields
